Split reauth tests in plex
