+++ tests/components/plex/test_init.py
@@ -15,7 +15,7 @@ from homeassistant.components.plex.models import (
     TRANSIENT_SECTION,
     UNKNOWN_SECTION,
 )
-from homeassistant.config_entries import ConfigEntryState
+from homeassistant.config_entries import SOURCE_REAUTH, ConfigEntryState
 from homeassistant.const import (
     CONF_TOKEN,
     CONF_URL,
@@ -336,3 +336,27 @@ async def test_setup_with_limited_credentials(
 
     assert len(hass.config_entries.async_entries(const.DOMAIN)) == 1
     assert entry.state is ConfigEntryState.LOADED
+
+
+async def test_trigger_reauth(
+    hass: HomeAssistant,
+    entry: MockConfigEntry,
+    mock_plex_server,
+    mock_websocket,
+) -> None:
+    """Test setup and reauthorization of a Plex token."""
+
+    assert entry.state is ConfigEntryState.LOADED
+
+    with patch(
+        "plexapi.server.PlexServer.clients", side_effect=plexapi.exceptions.Unauthorized
+    ), patch("plexapi.server.PlexServer", side_effect=plexapi.exceptions.Unauthorized):
+        trigger_plex_update(mock_websocket)
+        await wait_for_debouncer(hass)
+
+    assert len(hass.config_entries.async_entries(const.DOMAIN)) == 1
+    assert entry.state is not ConfigEntryState.LOADED
+
+    flows = hass.config_entries.flow.async_progress()
+    assert len(flows) == 1
+    assert flows[0]["context"]["source"] == SOURCE_REAUTH
